external/telegram_sheets_bot/requirements.txt
# Loyiha bog'liqliklari
aiogram>=3.4,<4
aiohttp>=3.9,<4
httpx>=0.27
python-dotenv>=1.0
gspread>=6.0
google-auth>=2.29
# Bepul ovoz (lokal Whisper; OPENAI kaliti shart emas)
faster-whisper>=1.0.0
openpyxl>=3.1.0
